backend/foodgram/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-colorfield==0.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.0.1
drf-extra-fields==3.4.1
gunicorn==20.0.4
Pillow==9.4.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
pytz==2020.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.3.1